moved FileRetrieverWebSocketClient to web_socket module.
